refactored to move Grid factory functionality into Grid's constructor
refactored to move Grid factory functionality into Grid's constructor
